+++ frontend/src/components/ui/editor/nodes/nodes.ts
@@ -17,6 +17,7 @@ import { KeywordNode } from "@/components/ui/editor/nodes/keyword-node";
 import { LayoutContainerNode } from "@/components/ui/editor/nodes/layout-container-node";
 import { LayoutItemNode } from "@/components/ui/editor/nodes/layout-item-node";
 import { MentionNode } from "@/components/ui/editor/nodes/mention-node";
+import { WikilinkNode } from "@/components/ui/editor/nodes/wikilink-node";
 
 export const nodes: ReadonlyArray<Klass<LexicalNode> | LexicalNodeReplacement> = [
   HeadingNode,
@@ -44,4 +45,5 @@ export const nodes: ReadonlyArray<Klass<LexicalNode> | LexicalNodeReplacement> =
   TweetNode,
   YouTubeNode,
   AutocompleteNode,
+  WikilinkNode,
 ];